Refactor network monitoring functionality and improve screen monitor logging
