+++ services/peripherals/app/src/screen_monitor.py
@@ -1,54 +1,4 @@
-import speedtest
 import subprocess
-
-class NetworkMonitor:
-    def __init__(self, event_dispatcher):
-        self.servers = []
-        self.threads = None
-        self.speed_test = speedtest.Speedtest()
-        self.dispatcher = event_dispatcher
-    
-    def get_servers(self):
-        self.speed_test.get_servers(self.servers)
-    
-    def get_best_server(self):
-        self.speed_test.get_best_server()
-    
-    def download(self):
-        self.speed_test.download(threads=self.threads)
-    
-    def upload(self):
-        self.speed_test.upload(threads=self.threads)
-    
-    def run_speed_test(self):
-        self.get_servers()
-        self.get_best_server()
-        self.download()
-        self.upload()
-    
-    def get_results(self):
-        return self.speed_test.results.dict()
-    
-    def check_internet_speed(self):
-        self.run_speed_test()
-        results = self.get_results()
-        if results:
-            print(f"Download = {results['download']/1000000}Mbps and upload = {results['upload']/1000000}Mbps")
-            self.dispatcher.dispatch_event("send_network_speed", results)
-    
-    def check_internet_connection(self):
-        ps = subprocess.Popen(['iwgetid'], stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-        try:
-            output = subprocess.check_output(('grep', 'ESSID'), stdin=ps.stdout)
-            if 'TP-Link_A58E' in str(output):
-                print("Connected to the TP-Link_A58E network")
-                self.dispatcher.dispatch_event("send_network_status", "connected")
-            else:
-                print(f"Could not find the TP-Link_A5BE network in the output: {str(output)}")
-        except subprocess.CalledProcessError:
-            # grep did not match any lines
-            print("No wireless networks connected")
-
 import time
 
 class ScreenMonitor:
@@ -74,7 +24,7 @@ class ScreenMonitor:
 
     def _configure_sleep_timer(self, payload):
         configuration = payload.get("control", False)
-        # print(f"sleep timer has been configured to: {configuration}")
+        print(f"sleep timer has been configured to: {configuration}")
         self.is_sleep_timer_enabled = configuration
         self._wake_up_screen()
 
@@ -101,11 +51,11 @@ class ScreenMonitor:
             return
         
         try:
-            subprocess.run(
-                f'echo {self.screen_dim_value} | sudo tee /sys/class/backlight/4-0045/brightness',
-                shell=True,
-                check=True
-            )
+            # subprocess.run(
+            #     f'echo {self.screen_dim_value} | sudo tee /sys/class/backlight/4-0045/brightness',
+            #     shell=True,
+            #     check=True
+            # )
             pass
             # Return a success response
         except subprocess.CalledProcessError as e:
@@ -121,11 +71,11 @@ class ScreenMonitor:
         self.dispatcher.dispatch_event("send_screen_status", "awake")
         for brightness in range(0, self.brightness):            
             try:
-                subprocess.run(
-                    f'echo {brightness} | sudo tee /sys/class/backlight/4-0045/brightness',
-                    shell=True,
-                    check=True
-                )
+                # subprocess.run(
+                #     f'echo {brightness} | sudo tee /sys/class/backlight/4-0045/brightness',
+                #     shell=True,
+                #     check=True
+                # )
                 pass
             except subprocess.CalledProcessError as e:
                 # Return an error response if the command fails
